Fix minor warnings found with a different compiler

+++ coap/block1.c
@@ -125,7 +125,7 @@ uint8_t coap_block1_handler(lwm2m_block1_data_t ** pBlock1Data,
 
     if (blockMore)
     {
-        *outputLength = -1;
+        *outputLength = (size_t)-1;
         return COAP_231_CONTINUE;
     }
     else
